requirements.txt
matplotlib==3.5.3
numpy==1.21.6
opencv_python==4.6.0.66
Pillow==9.5.0
scipy==1.7.3
